config security + jwt - update
